+++ man-roxygen/ClimateNA_dataPath.R
@@ -1,0 +1,3 @@
+#' @param dataPath character. path to top-level directory for reading
+#'                 and writing ClimateNA data.
+