+++ man-roxygen/ClimateNA_period.R
@@ -1,0 +1,2 @@
+#' @param period character. climate normal periods to use (e.g., `c("1951_1980", "1981_2010")`).
+