+++ man-roxygen/ClimateNA_type.R
@@ -1,0 +1,4 @@
+#' @param type character. denotes the table in the sqlite database to access.
+#'             one of 'historic', 'future', 'historic_normals', or 'future_normals'.
+#'
+