+++ man-roxygen/ClimateNA_years.R
@@ -1,0 +1,2 @@
+#' @param years integer. consecutive data years to use (e.g., `2011:2100`).
+