+++ tests/testthat/test-makeMDC.R
@@ -6,16 +6,19 @@ test_that("makeMDC works", {
   skip_if_not(interactive())
 
   dPath <- file.path(tempdir(), "test_makeMDC")
-  historicalClimatePath <- reproducible::checkPath(file.path(dPath, "climate", "historic"), create = TRUE)
-  historicalClimateArchive <- file.path(historicalClimatePath, "Alberta.zip")
-
-  gid <- googledrive::as_id("1we9GqEVAORWLbHi3it66VnCcvLu85QIk") ## Alberta.zip
-  tbl <- googledrive::drive_download(file = gid, path = historicalClimateArchive)
-
-  archive::archive_extract(historicalClimateArchive, historicalClimatePath)
-
+  climateType <- "historic"
+  climatePath <- reproducible::checkPath(file.path(dPath, "climate", climateType), create = TRUE)
+  tileIDs <- c(16, 17, 26, 27)
   yrs <- c(2011:2015)
-  mdc <- makeMDC(tools::file_path_sans_ext(historicalClimateArchive), years = yrs, droughtMonths = 4:9)
+
+  tileURLs <- getClimateURLs(type = climateType, tiles = tileIDs, years = yrs, msy = "M")
+
+  getClimateTiles(tileIDs, tileURLs, climatePath)
+  allDirs <- file.path(climatePath, rep(tileIDs, length(yrs)), paste0("Year_", yrs, "M")) |> sort()
+  expect_true(all(dir.exists(allDirs)))
+
+  mdc <- makeMDC(srcdir = climatePath, dstdir = dPath,
+                 tiles = tileIDs, years = yrs, droughtMonths = 4:9, cl = NULL)
 
   expect_s4_class(mdc, "SpatRaster")
   expect_identical(names(mdc), paste0("mdc", yrs))