+++ NEWS.md
@@ -1,3 +1,8 @@
+# climateData 2.0.0
+
+- improved ClimateNA workflow to use raster tiles instead of individual provinces, and use GDAL VRTs for creating output rasters;
+- updated climate data sources generated using ClimateNA v7.4.2 (released Aug 2023). As of ClimateNA v7.41, climate variables with decimals are no longer converted to integers in raster format.
+
 # climateData 1.0.0
 
 ## Dependency changes